+++ test/cli_test.py
@@ -2,6 +2,7 @@
 import json
 import os
 import pytest
+import re
 import sys
 import tempfile
 import traceback
@@ -11,11 +12,14 @@ from unittest import mock
 import click
 import click.testing
 import pytest_asyncio
+import toml
 
 from modal import Client
 from modal.cli.entry_point import entrypoint_cli
 from modal_proto import api_pb2
 from modal_utils.async_utils import asyncnullcontext
+
+from .supports.skip import skip_windows
 
 dummy_app_file = """
 import modal
@@ -109,13 +113,15 @@ def test_secret_list(servicer, set_env_client):
 
 
 def test_app_token_new(servicer, set_env_client, server_url_env, modal_config):
-    with modal_config():
+    with modal_config() as config_file_path:
         _run(["token", "new", "--profile", "_test"])
+        assert "_test" in toml.load(config_file_path)
 
 
 def test_app_setup(servicer, set_env_client, server_url_env, modal_config):
-    with modal_config():
+    with modal_config() as config_file_path:
         _run(["setup", "--profile", "_test"])
+        assert "_test" in toml.load(config_file_path)
 
 
 def test_run(servicer, set_env_client, test_dir):
@@ -336,35 +342,43 @@ def mock_shell_pty():
             env_term_program=os.environ.get("TERM_PROGRAM"),
         )
 
+    captured_out = []
+
+    async def write_to_fd(fd: int, data: bytes):
+        nonlocal captured_out
+        captured_out.append((fd, data))
+
     with mock.patch("rich.console.Console.is_terminal", True), mock.patch(
         "modal._pty.get_pty_info", mock_get_pty_info
-    ), mock.patch("modal._pty.write_stdin_to_pty_stream", asyncnullcontext):
-        yield
+    ), mock.patch("modal._container_exec.get_pty_info", mock_get_pty_info), mock.patch(
+        "modal._pty.write_stdin_to_pty_stream", asyncnullcontext
+    ), mock.patch(
+        "modal._container_exec.handle_exec_input", asyncnullcontext
+    ), mock.patch(
+        "modal._container_exec._write_to_fd", write_to_fd
+    ):
+        yield captured_out
 
 
-@pytest.mark.usefixtures("mock_shell_pty")
-def test_shell(servicer, set_env_client, test_dir):
+@skip_windows("modal shell is not supported on Windows.")
+def test_shell(servicer, set_env_client, test_dir, mock_shell_pty):
     stub_file = test_dir / "supports" / "app_run_tests" / "default_stub.py"
     webhook_stub_file = test_dir / "supports" / "app_run_tests" / "webhook.py"
 
-    ran_cmd = None
-
-    @servicer.function_body
-    def dummy_exec(cmd: str):
-        nonlocal ran_cmd
-        ran_cmd = cmd
-
     # Function is explicitly specified
     _run(["shell", stub_file.as_posix() + "::foo"])
-    assert ran_cmd == "/bin/bash"
+    assert mock_shell_pty == [(1, b"Hello World")]
+    mock_shell_pty.clear()
 
     # Function is explicitly specified
     _run(["shell", webhook_stub_file.as_posix() + "::foo"])
-    assert ran_cmd == "/bin/bash"
+    assert mock_shell_pty == [(1, b"Hello World")]
+    mock_shell_pty.clear()
 
     # Function must be inferred
     _run(["shell", webhook_stub_file.as_posix()])
-    assert ran_cmd == "/bin/bash"
+    assert mock_shell_pty == [(1, b"Hello World")]
+    mock_shell_pty.clear()
 
 
 def test_app_descriptions(servicer, server_url_env, test_dir):
@@ -531,10 +545,45 @@ def test_cls(servicer, set_env_client, test_dir):
     _run(["run", f"{stub_file.as_posix()}::AParametrized.some_method", "--x", "42", "--y", "1000"])
 
 
-def test_profile_list(servicer, server_url_env):
-    res = _run(["profile", "list"])
-    assert "Profile" in res.stdout
-    assert "Workspace" in res.stdout
+def test_profile_list(servicer, server_url_env, modal_config):
+    config = """
+    [test-profile]
+    token_id = "ak-abc"
+    token_secret = "as-xyz"
 
-    res = _run(["profile", "list", "--json"])
-    json.loads(res.stdout)
+    [other-profile]
+    token_id = "ak-123"
+    token_secret = "as-789"
+    active = true
+    """
+
+    with modal_config(config):
+        res = _run(["profile", "list"])
+        table_rows = res.stdout.split("\n")
+        assert re.search("Profile .+ Workspace", table_rows[1])
+        assert re.search("test-profile .+ test-username", table_rows[3])
+        assert re.search("other-profile .+ test-username", table_rows[4])
+
+        res = _run(["profile", "list", "--json"])
+        json_data = json.loads(res.stdout)
+        assert json_data[0]["name"] == "test-profile"
+        assert json_data[0]["workspace"] == "test-username"
+        assert json_data[1]["name"] == "other-profile"
+        assert json_data[1]["workspace"] == "test-username"
+
+        orig_env_token_id = os.environ.get("MODAL_TOKEN_ID")
+        orig_env_token_secret = os.environ.get("MODAL_TOKEN_SECRET")
+        os.environ["MODAL_TOKEN_ID"] = "ak-abc"
+        os.environ["MODAL_TOKEN_SECRET"] = "as-xyz"
+        try:
+            res = _run(["profile", "list"])
+            assert "Using test-username workspace based on environment variables" in res.stdout
+        finally:
+            if orig_env_token_id:
+                os.environ["MODAL_TOKEN_ID"] = orig_env_token_id
+            else:
+                del os.environ["MODAL_TOKEN_ID"]
+            if orig_env_token_secret:
+                os.environ["MODAL_TOKEN_SECRET"] = orig_env_token_secret
+            else:
+                del os.environ["MODAL_TOKEN_SECRET"]