+++ test/gpu_test.py
@@ -142,25 +142,13 @@ def test_a100_20gb_gpu_unsupported():
         stub.function(gpu=modal.gpu.A100(memory=20))(dummy)
 
 
-A100_GPU_COUNT_MAPPING = {1: api_pb2.GPU_TYPE_A100, **{i: api_pb2.GPU_TYPE_A100 for i in range(2, 5)}}
-
-
-@pytest.mark.parametrize("count,gpu_type", A100_GPU_COUNT_MAPPING.items())
-def test_gpu_type_selection_from_count(client, servicer, count, gpu_type):
+@pytest.mark.parametrize("count", [1, 2, 3, 4])
+def test_gpu_type_selection_from_count(client, servicer, count):
     import modal
 
     stub = Stub()
 
-    # Functions that use A100 20GB can only request one GPU
-    # at a time.
-    with pytest.raises(ValueError):
-        with pytest.warns(DeprecationError):
-            stub.function(gpu=modal.gpu.A100(count=2, memory=20))(dummy)
-        with stub.run(client=client):
-            pass
-
-    # Task type changes whenever user asks more than 1 GPU on
-    # an A100.
+    # Task type does not change when user asks more than 1 GPU on an A100.
     stub.function(gpu=modal.gpu.A100(count=count))(dummy)
     with stub.run(client=client):
         pass
@@ -168,4 +156,4 @@ def test_gpu_type_selection_from_count(client, servicer, count, gpu_type):
     func_def = next(iter(servicer.app_functions.values()))
 
     assert func_def.resources.gpu_config.count == count
-    assert func_def.resources.gpu_config.type == gpu_type
+    assert func_def.resources.gpu_config.type == api_pb2.GPU_TYPE_A100